% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blankODs}
\alias{blankODs}
\title{Blank optical density data}
\usage{
blankODs(
  data,
  method = "fromPlate",
  perTimePoint = TRUE,
  groups = NULL,
  values = NULL,
  tukeyK = NULL
)
}
\arguments{
\item{data}{A tibble containing OD data as produced by the function \code{processODdata}.}

\item{method}{Determines whether to blank using OD values on the same plate (\code{method} = "fromPlate"),
or fixed OD values as specified in \code{blankValues} argument (\code{method} = "fixed").
For more information, see Details below.}

\item{perTimePoint}{Determines whether blanking will be performed on a per-time-point basis (\code{perTimePoint} = "TRUE"),
or whether blank OD values are averaged through time before being subtracted (\code{perTimePoint} = "FALSE").}

\item{groups}{One or several columns in the \code{data} tibble by which blanking should be grouped.
This argument is only used when \code{method} = "fromPlate". See Details below.}

\item{values}{OD value(s) for blanking. This argument is only used when \code{method} = "valuesProvided".
See Details below.}

\item{tukeyK}{This is an argument for calculating the blanked ODs.
When averaging OD of BLANK wells (across wells and/or across time), outliers may be excluded.
Here, Tukey's fences method is used, excluding all OD values that are more than tukeyK times the
interquartile range away from the upper and lower quartile, respectively.
When kTukey is NULL (the default), no outliers will be excluded.}
}
\value{
The original \code{data} tibble with an additional column \code{blankedOD}.
}
\description{
\code{blankODs} takes a data file (tibble) of optical density (OD) data and uses one of several methods to blank the OD data.
}
\details{
Currently, two methods for blanking are implemented. When \code{method} = "fromPlate" (the default),
OD values from wells on the same plate that are designated as type "BLANK" will be used for blanking.
If no groups are provided (argument \code{groups = NULL}, the default), OD values of all BLANK wells
will be averaged, and the blanked OD will then be calculated as the original OD in each DATA well,
minus the this average value. However, it is also possible to group the blanked values by one or more
of the variables in the data tibble, and then subtract the group average from the corresponding DATA well ODs.
For examples, if there is a variable 'Medium' in the tibble, then with \code{blankGroups = 'Medium'},
averages for blanking will be taken across all wells with \code{wellType="BLANK"} for each value of this column
(e.g. "LB", "M9" etc.), and subtracted from OD for data wells with the same 'Medium' values.

When \code{method} = "fixed", one or several blank OD values, supplied in the \code{blankValues} argument,
are subtracted from all corresponding DATA wells. In the simplest case, \code{blankValues} is just a single
number that is subtracted. However, \code{blankValues} can also be a data frame or tibble of values.
In this case, one of the columns needs to be named "blankOD" (the blank OD values), and additional columns
"Time_min" and/or columns for grouping. For example, if in the \code{data} tibble that is supplied there
is a variable "Medium" then the \code{blankValues} tibble could have a corresponding column "Medium"
specifying the different values for which blank ODs are provided.

The \code{perTimePoint} argument is applicable to both methods.
When \code{perTimePoint} = "TRUE" (the default), blanking will be performed on a per-time-point basis,
i.e. at each time point, the blanking value at this time point is subtracted from the respective wells to be blanked.
When \code{perTimePoint} = "FALSE", the OD of blank values is averaged across all time points (if there are several),
and then subtracted from ODs at all time points. Note that with this latter method,
any trend in OD in the BLANK wells will lead to spurious results because this trend will be averaged out.
}
