% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_for_CRfitting}
\alias{export_for_CRfitting}
\title{Export OD dataset for fitting consumer-resource models}
\usage{
export_for_CRfitting(
  ODs,
  inoculum_IDs = c("Plate", "Replicate", "Strain"),
  resource_IDs = c("Plate", "Replicate")
)
}
\arguments{
\item{ODs}{Data frame containing OD data, as produced by the processODData function}

\item{inoculum_IDs}{A character string specifying columns in the ODs data frame that determine
which wells were inoculated with the same culture.
For example, with the default value, \code{inoculum_IDs = c("Plate", "Replicate", "Strain")}, all wells with unique
combinations of values in the Plate, Replicate and Strain columns are assumed to be have been inoculated from the
same culture.}

\item{resource_IDs}{A character string specifying columns in the ODs data frame that determine
which wells have the identical growth medium.
For example, with the default value, \code{resource_IDs = c("Plate", "Replicate")}, all wells with unique
combinations of values in the Plate and Replicate columns are assumed to contain growth media that
is identical in terms of composition and batch,
but all media in the different wells within a single plate is assumed to be identical.}
}
\value{
A data frame of OD values,
with some rows filtered out and some additional columns added compared to the input data frame.
}
\description{
This function prepares an OD dataset for downstream fitting of consumer-resource models.
It takes as input a data frame with OD data, in tidy format as produced by the
processODData function. After some basic filtering (removing wells with \code{WellType == "EMPTY"}
and rows where either time or OD has NA values), it then adds four numeric ID columns to the data frame
that are essential for the fitting procedure:
\itemize{
\item Plate_ID: A unique number for each individual plate (i.e., each unique combination of \code{PlateID} and \code{Replicate})
\item Timeseries_ID: A unique number for each time series (i.e. well) within a plate
\item Inoculum_ID: A unique number for each independent inoculum
\item Resource_ID: A unique number for each independently prepared growth medium
}
}
\details{
Whilst \code{Plate_ID} and \code{Timeseries_ID} are determined automatically from default columns,
correct assignment of \code{Inoculum_ID} and \code{Resource_ID} may require the user to either adjust the function arguments
\code{inoculum_IDs} and \code{resource_IDs}, or, in more complex scenarios, to change the output data frame after function call.
}
