% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{getGrowthParameters}
\alias{getGrowthParameters}
\title{Estimate growth parameters for a single vector of OD measurements}
\usage{
getGrowthParameters(times, ODs, method = "easylinear", ...)
}
\arguments{
\item{times}{Time points where OD measurements have been taken}

\item{ODs}{OD measurements, ideally blanked}

\item{method}{Currently, only the default value "easylinear" is supported}

\item{...}{Further arguments to be passed on to the method used}
}
\value{
A vector of estimated growth parameters
}
\description{
Estimate growth parameters for a single vector of OD measurements
}
