% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{importODFile}
\alias{importODFile}
\title{Import raw OD data from file}
\usage{
importODFile(fileName)
}
\arguments{
\item{fileName}{The name (including path) of the file. Supported file types are xlsx, csv and tsv.}
}
\value{
An (untidy) tibble in which each row represents one time point and each column represents one well, in addition to columns for date, plate reader model, time and temperature.
}
\description{
This function imports a single raw data file produced by a microplate reader. Currently supported are BioTek Synergy HT and Epoch plate readers.
}
