% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotODs}
\alias{plotODs}
\title{Plot OD data through time}
\usage{
plotODs(
  data,
  replicates = NULL,
  blanked = TRUE,
  border = TRUE,
  smallScreen = FALSE
)
}
\arguments{
\item{data}{The data to be plotted, as produced by the \code{processODData} function.}

\item{replicates}{A vector indicating which replicates should be plotted.}

\item{blanked}{A boolean value indicating whether the blanked OD should be plotted or not.}

\item{border}{A boolean value indicating whether or not to include the plate border (rows A and H, columns 1 and 12) in the plot.}

\item{smallScreen}{If \code{TRUE}, many of the details in the plot (legend, axes labels etc.) will be omitted so as to fit all the data on a small screen.}
}
\value{
A ggplot object.
}
\description{
This function produces a plot where OD through time is shown for each well of a 96-well plate.
}
