% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcBlanks}
\alias{qcBlanks}
\title{Obtain quality control data and plots on blank wells}
\usage{
qcBlanks(data, blankGroups = NULL)
}
\arguments{
\item{data}{The data set, as produced by processODData}

\item{blankGroups}{If specified, one or several columns in the \code{data} tibble by which blanking should be grouped. For examples, if there is a variable 'Medium' in the tibble, then with \code{blankGroups = 'Medium'}, averages for blanking will be taken across all wells with \code{wellType="BLANK"} for each value of this column (e.g. "LB", "M9" etc.), and subtracted from OD for data wells with the same Medium values.}
}
\value{
A list of two elements: a summary table and a list of plots for each plate. If there are no blanks at all in the data set, \code{NULL} is returned.
}
\description{
Obtain quality control data and plots on blank wells
}
\keyword{internal}
