% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcODData}
\alias{qcODData}
\title{Perform quality control on OD data set}
\usage{
qcODData(data, blankGroups = NULL, path = ".")
}
\arguments{
\item{data}{The data set, as produced by processODData}

\item{blankGroups}{If specified, one or several columns in the \code{data} tibble by which blanking should be grouped. For examples, if there is a variable 'Medium' in the tibble, then with \code{blankGroups = 'Medium'}, averages for blanking will be taken across all wells with \code{wellType="BLANK"} for each value of this column (e.g. "LB", "M9" etc.), and subtracted from OD for data wells with the same Medium values.}

\item{path}{The path where to save the pdf report. Defaults to the current working directory.}
}
\value{
A list containing various statistics and plots for quality control.
}
\description{
This function performs a quality control analysis of an OD dataset,
covering both temperature through time and OD for wells designated as blanks.
The report is saved as a pdf file but the data and plots are also directly returned to the user.
}
