% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcTemperature}
\alias{qcTemperature}
\title{Obtain quality control data and plots on temperature}
\usage{
qcTemperature(data)
}
\arguments{
\item{data}{The data set, as produced by processODData}
}
\value{
A list of two elements: a summary table and a list of plots for each plate
}
\description{
Obtain quality control data and plots on temperature
}
\keyword{internal}
