% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{shinyPlate}
\alias{shinyPlate}
\title{A shiny app to explore OD data from a 96-well plate.}
\usage{
shinyPlate(data)
}
\arguments{
\item{data}{The data to be plotted, as produced by the \code{processODData} function.}
}
\value{
Starts a shiny app.
}
\description{
This shiny app lets you choose replicates and a number of options to plot OD data through time for each well of a 96-well plate.
}
