% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{specPlot_fullFact}
\alias{specPlot_fullFact}
\title{Generate a pdf file with a plot of a plate design}
\usage{
specPlot_fullFact(
  plateName,
  replicate,
  vWellType,
  rows,
  columns,
  border,
  fileName
)
}
\arguments{
\item{plateName}{Name of the plate.}

\item{replicate}{Replicate plate number.}

\item{vWellType}{A matrix specifying the type of each well ("DATA", "BLANK" or "EMPTY")}

\item{rows}{A vector of values of the row variable}

\item{columns}{A vector of values of the column variable}

\item{border}{Type of border}

\item{fileName}{File name of for the pdf (including path)}
}
\description{
Generate a pdf file with a plot of a plate design
}
