% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{specPlot_wrapping}
\alias{specPlot_wrapping}
\title{Generate a pdf file with a plot of a wrapping plate design}
\usage{
specPlot_wrapping(
  plateName,
  replicate,
  vWellType,
  vVarWrap,
  border,
  fileName,
  vVarGroup = NULL
)
}
\arguments{
\item{plateName}{Name of the plate.}

\item{replicate}{Replicate plate number.}

\item{vWellType}{A matrix specifying the type of each well ("DATA", "BLANK" or "EMPTY")}

\item{border}{Type of border}

\item{fileName}{File name of for the pdf (including path)}

\item{rows}{A vector of values of the row variable}

\item{columns}{A vector of values of the column variable}
}
\description{
Generate a pdf file with a plot of a wrapping plate design
}
\keyword{internal}
