% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyseODData}
\alias{analyseODData}
\title{Title}
\usage{
analyseODData(
  data,
  method = "easylinear",
  tmin = 0,
  tmax = Inf,
  r2Threshold = 0.5,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
