% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{blankODs}
\alias{blankODs}
\title{Blank optical density data}
\usage{
blankODs(data, groups = NULL)
}
\arguments{
\item{data}{A tibble containing OD data as produced by the function \code{processODdata}.}

\item{groups}{If specified, one or several columns in the \code{data} tibble by which blanking should be grouped. For examples, if there is a variable 'Medium' in the tibble, then with \code{blankGroups = 'Medium'}, averages for blanking will be taken across all wells with \code{wellType="BLANK"} for each value of this column (e.g. "LB", "M9" etc.), and subtracted from OD for data wells with the same Medium values.}
}
\value{
The original \code{data} tibble with an additional column \code{blankedOD}.
}
\description{
\code{blankODs} takes a data file (tibble) of optical density (OD) data and, for each individual plate and time step, uses the average of wells designated as \code{wellType="BLANK"} to blank all the data cells (\code{wellType="BLANK"}).
}
\keyword{internal}
